Uploads

+++ src/main/java/hello/DeleteFilesThread.java
@@ -5,21 +5,38 @@ import java.io.File;
 public class DeleteFilesThread extends Thread {
 
     String a,b;
+    int n;
 
-    DeleteFilesThread(String s1, String s2) {
+
+    DeleteFilesThread(String s1, String s2, int number) {
         a = s1;
         b = s2;
+        n = number;
     }
 
-    public void run(){
-        try {
-            sleep(80000);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
+    public void run() {
+
+        if (n == 2) {
+
+            try {
+                sleep(80000);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+            File file = new File(a);
+            file.delete();
+            file = new File(b);
+            file.delete();
         }
-        File file = new File(a);
-        file.delete();
-        file = new File(b);
-        file.delete();
+
+        if (n == 1) {
+            try {
+                sleep(80000);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+            File file = new File(a);
+            file.delete();
+        }
     }
 }
